gcp-finops/scripts/cloudfunctions/recommender-checker/requirements.txt
google-cloud-asset==3.9.0
google-cloud-recommender==2.7.3
